refactor: simplify backend read prducts challenge
