+++ sessions/backend-read/products/pages/api/products/[id].js
@@ -6,7 +6,8 @@ export default function handler(request, response) {
   const product = products.find((product) => product.id === id);
 
   if (!product) {
-    return response.status(404).json({ status: "Not Found" });
+    response.status(404).json({ status: "Not Found" });
+    return;
   }
 
   response.status(200).json(product);
